requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.2
django-heroku==0.3.1
django-on-heroku==1.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
numpy==1.23.4
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
whitenoise==6.2.0
